Initial Commit for PokeMap
Thanks to Google's April Fools, we took it upon ourselves to build an organizer for all the pokemon in the new Google Maps mini game but  creating a centralized list of Pokemon and their respective locations. This started as an exercise in quick application develop as a team and ended up being a fun little project.

Provides everything needed to track, add, and  catch all 150 Pokemon convienently from a smart phone through the use of Google Maps query strings to take a user to a location close to the designated Pokemon. The option to add newer/more accurate locations is provided as well.

All listings started out blank and if they're not provided for, the option to add their name and locations is opened up.

All calls are AJAX'd to a SQL db that had 150 rows preloaded and was filled by test users.

Front end is handled with AngularJS and Boostrap 3.

Backend is straight PHP/SQL.

Developed by myself and Christopher White (christopherwhite200@gmail.com)

See it working live at http://pokemap.info

**Php left out for database security concerns**